Move timsort to the tests directory
